Customize the handling of the username #12
When a user logs in via OIDC, a separate, unique user ID is transmitted
by the OIDC provider. This user ID is now used as a username (and at
the same time as a unique ID).

The username used before is actually more the display name. It is now
used as "PreferredUsername".